+++ R/get_correlations.R
@@ -9,6 +9,8 @@ setMethod(
   f = "get_correlations",
   signature = c(x = "PCA"),
   definition = function(x, sup_name = ".sup") {
+    arkhe::assert_scalar(sup_name, "character")
+
     corr <- x@columns@principal / x@columns@distances
     suppl <- x@columns@supplement