+++ R/get_cos2.R
@@ -9,8 +9,9 @@ setMethod(
   f = "get_cos2",
   signature = c(x = "MultivariateAnalysis"),
   definition = function(x, margin = 1, sup_name = ".sup") {
+    arkhe::assert_scalar(margin, "numeric")
+    arkhe::assert_scalar(sup_name, "character")
 
-    margin <- margin[[1L]]
     if (margin == 1) {
       cos2 <- x@rows@cosine
       suppl <- x@rows@supplement