+++ R/get_distances.R
@@ -9,8 +9,8 @@ setMethod(
   f = "get_distances",
   signature = c(x = "MultivariateAnalysis"),
   definition = function(x, margin = 1) {
+    arkhe::assert_scalar(margin, "numeric")
 
-    margin <- margin[[1L]]
     if (margin == 1) {
       d2 <- x@rows@distances
       names(d2) <- x@rows@names