+++ R/pca.R
@@ -40,11 +40,6 @@ setMethod(
   definition = function(object, center = TRUE, scale = TRUE, rank = NULL,
                         sup_row = NULL, sup_col = NULL,
                         weight_row = NULL, weight_col = NULL) {
-    ## Check missing values
-    if (anyNA(object)) stop("Missing values detected.", call. = FALSE)
-    ## Check dimensions
-    if (any(dim(object) == 0)) stop("Empty matrix.", call. = FALSE)
-
     # Fix dimension names
     names_row <- rownames(object)
     names_col <- colnames(object)
@@ -55,6 +50,12 @@ setMethod(
     is_row_sup <- find_variable(sup_row, nrow(object), names = rownames(object))
     is_col_sup <- find_variable(sup_col, ncol(object), names = colnames(object))
     N <- object[!is_row_sup, !is_col_sup, drop = FALSE]
+
+    ## Check missing values
+    arkhe::assert_missing(N)
+
+    ## Check dimensions
+    arkhe::assert_filled(N)
 
     # Dimension of the solution
     ndim <- min(rank, dim(N) - 1)