+++ R/wrap_hull.R
@@ -9,6 +9,11 @@ setMethod(
   f = "wrap_hull",
   signature = c(x = "MultivariateAnalysis"),
   definition = function(x, margin = 1, axes = c(1, 2), group = NULL) {
+    ## Validation
+    arkhe::assert_scalar(margin, "numeric")
+    arkhe::assert_type(axes, "numeric")
+    arkhe::assert_length(axes, 2)
+
     ## Get coordinates
     data <- get_coordinates(x, margin = margin)
     data <- data[, axes]